feat: Refactor ApplicationsTable for improved state management and organization
- Extracted application state management logic into a new custom hook, useApplicationState, to enhance separation of concerns and maintainability.
- Introduced a new hook, useTableConfig, to manage table configuration and state, streamlining the ApplicationsTable component.
- Added a useTableSearch hook for improved search functionality, encapsulating fuzzy and global search logic.
- Cleaned up the ApplicationsTable component by removing inline state management and filtering logic, resulting in clearer and more organized code.
